+++ src/ska_dishln_pointing_device/commands/stop_program_track_table.py
@@ -38,7 +38,9 @@ class StopProgramTrackTable(FastCommand):
                 self.component_manager.track_table_thread
                 and self.component_manager.track_table_thread.is_alive()
             ):
-                self.component_manager.track_table_thread.join()
+                # The track_table_thread is joined with a short timeout to
+                # prevent the Stop command from triggering a Tango timeout
+                self.component_manager.track_table_thread.join(timeout=2.8)
             self.component_manager.update_pointing_program_track_table([])
             self.component_manager.current_mapping_scan_obj = None
             self.logger.info(